+++ hardware/1Wire/1WireSystem.h
@@ -18,7 +18,7 @@ public:
    virtual float GetPressure(const _t1WireDevice& device) const=0;
    virtual bool GetLightState(const _t1WireDevice& device,int unit) const=0;
    virtual unsigned int GetNbChannels(const _t1WireDevice& device) const=0;
-   virtual uint32_t GetCounter(const _t1WireDevice& device,int unit) const=0;
+   virtual unsigned long GetCounter(const _t1WireDevice& device,int unit) const=0;
    virtual int GetVoltage(const _t1WireDevice& device,int unit) const=0;
    virtual float GetIlluminance(const _t1WireDevice& device) const=0;
 };
