+++ hardware/ETH8020.cpp
@@ -211,7 +211,7 @@ void CETH8020::GetMeterDetails()
 	}
 	size_t ii;
 	std::string tmpstr;
-	size_t pos1;
+	int pos1;
 	int Idx = 0;
 	for (ii = 1; ii < results.size(); ii++)
 	{
