+++ hardware/EnOceanESP3.h
@@ -32,7 +32,7 @@ public:
     ~CEnOceanESP3();
 	bool WriteToHardware(const char *pdata, const unsigned char length);
 	void SendDimmerTeachIn(const char *pdata, const unsigned char length);
-	uint32_t m_id_base;
+	unsigned long m_id_base;
 private:
 	void Init();
 	bool StartHardware();
