+++ hardware/MySensorsBase.cpp
@@ -1871,7 +1871,7 @@ void MySensorsBase::ParseLine()
 		{
 			if (!payload.empty())
 			{
-				uint64_t idx = boost::lexical_cast<uint64_t>(payload);
+				unsigned long long idx = boost::lexical_cast<unsigned long long>(payload);
 				int nValue;
 				std::string sValue;
 				if (m_mainworker.GetSensorData(idx, nValue, sValue))
